feat: implement isStaged function to check if a file is staged in git; add corresponding tests

+++ packer.js
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/no-require-imports */
 /* eslint-disable no-useless-escape */
 const { spawn } = require('child_process');
 const fs = require('fs-extra');
@@ -166,7 +167,9 @@ function bundleWithNpm() {
   if (!fs.existsSync(tgz)) {
     const filename2 = slugifyPkgName(`${packagejson.name}-${packagejson.version}.tgz`);
     const origintgz = join(__dirname, filename2);
-    fs.renameSync(origintgz, tgz);
+    if (fs.existsSync(origintgz) && origintgz !== tgz) {
+      fs.renameSync(origintgz, tgz);
+    }
   }
   const tgzlatest = join(releaseDir, slugifyPkgName(`${packagejson.name}.tgz`));
 
@@ -219,11 +222,9 @@ function parseVersion(versionString) {
  * create release/readme.md
  */
 async function addReadMe() {
-  if (['git-command-helper', 'cross-spawn'].includes(packagejson.name)) {
-    console.error('cannot run add readme on', packagejson.name);
-    return;
-  }
-  const { async: spawnAsync } = await import('cross-spawn');
+  const isCrossSpawn = packagejson.name == 'cross-spawn';
+  const isGitCommandHelper = packagejson.name == 'git-command-helper';
+  const { async: spawnAsync } = isCrossSpawn ? await import('./dist/index.js') : await import('cross-spawn');
   // set username and email on CI
   if (_isCI) {
     await spawnAsync('git', ['config', '--global', 'user.name', 'dimaslanjaka'], {
@@ -239,9 +240,11 @@ async function addReadMe() {
   /**
    * @type {typeof import('git-command-helper')}
    */
-  const gch = packagejson.name !== 'git-command-helper' ? require('git-command-helper') : require('./dist');
+  const { gitCommandHelper: gch } = isGitCommandHelper
+    ? await import('./dist/index.js')
+    : await import('git-command-helper');
 
-  const git = new gch.default(__dirname);
+  const git = new gch(__dirname);
   const branch = (await git.getbranch()).filter((o) => o.active)[0].branch;
   const gitlatest = await git.latestCommit();
